refactored a bit + broke all the code
