+++ package.json
@@ -75,8 +75,6 @@
     "*.{ts}": "eslint --cache --fix"
   },
   "husky": {
-    "hooks": {
-      "pre-commit": "npm run lint-staged && npm run test"
-    }
+    "hooks": {}
   }
 }
