+++ src/bin/serve.ts
@@ -1,6 +1,4 @@
-import fs from "fs"
-import url from "url"
-import http, { IncomingMessage, ServerResponse } from "http"
+import http from "http"
 import path from "path"
 import WebSocket from "ws"
 import chokidar from "chokidar"
@@ -9,68 +7,11 @@ import portfinder from "portfinder"
 import betterOpen from "better-opn"
 
 import build from "./build"
-import mimeType from "../utils/mimeTypes"
 import removeDirectories from "../fs/rmdir"
 import globalInfo from "../globalInfo"
 import linePrint from "../utils/linePrint"
 
-const configs = globalInfo.configs
-
-const createServer = (req: IncomingMessage, res: ServerResponse) => {
-  try {
-    // console.log(`${req.method} ${req.url}`)
-
-    const parsedUrl = url.parse(req.url!)
-    const sanitizePath = path
-      .normalize(parsedUrl.pathname!)
-      .replace(/^(\.\.[/\\])+/, "")
-
-    let pathname = path.join(process.cwd(), configs.buildPath, sanitizePath)
-
-    if (!fs.existsSync(pathname)) {
-      res.statusCode = 404
-      res.end(`File ${pathname} not found!`)
-      return
-    }
-
-    if (fs.statSync(pathname).isDirectory()) {
-      pathname = path.join(pathname, "index.html")
-    }
-
-    let data = fs.readFileSync(pathname)
-    const ext = path.parse(pathname).ext
-    res.setHeader("Content-type", mimeType[ext] || "text/plain")
-
-    if (ext === ".html") {
-      // @ts-ignore
-      data += `
-        <script>
-        if ('WebSocket' in window) {
-            const protocol = window.location.protocol === 'http:' ? 'ws://' : 'wss://';
-            const address = protocol + window.location.host + window.location.pathname;
-            const ws = new WebSocket(address);
-            ws.addEventListener('message', e => {
-              if (e.data === 'refresh-css') {
-                const links = document.querySelectorAll("link[rel='stylesheet']");
-                links.forEach(link => {
-                  link.href += "";
-                })
-              }
-
-              if(e.data === "reload-page") location.reload()
-            });
-        }else {
-          console.warn("Your browser doesn't support websockets, so can't live reload")
-        }
-        </script>  
-        `
-    }
-    res.end(data)
-  } catch (e) {
-    res.statusCode = 500
-    res.end(e.message)
-  }
-}
+import { createServer } from "../utils/cli-utils"
 
 const liveServer = async (usersPort = 3000) => {
   const port = await portfinder.getPortPromise({ port: usersPort })
