+++ src/generate/generateUlka.ts
@@ -4,7 +4,7 @@ import path from "path"
 import mkdir from "../fs/mkdir"
 import globalInfo from "../globalInfo"
 import parseUlka from "../parse/parseUlka"
-import absolutePath from "../utils/absolutePath"
+import { absolutePath } from "../utils/path-utils"
 import data from "./data"
 
 const configs = globalInfo.configs
