+++ src/utils/path-utils/index.ts
@@ -1,0 +1,2 @@
+export { default as absolutePath } from "./absolute-path"
+export { default as dataFromPath } from "./data-from-path"
